norm and bug
