+++ srcs/srcs_pars/free_tokens_and_ast.c
@@ -6,7 +6,7 @@
 /*   By: emagnani <emagnani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/20 18:20:35 by emagnani          #+#    #+#             */
-/*   Updated: 2025/01/20 15:05:42 by emagnani         ###   ########.fr       */
+/*   Updated: 2025/01/24 17:16:44 by emagnani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
