+++ srcs/srcs_pars/lexing_grammar.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   lexing_grammar.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kuru <kuru@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: emagnani <emagnani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/14 15:52:56 by emagnani          #+#    #+#             */
-/*   Updated: 2025/01/17 19:06:49 by kuru             ###   ########.fr       */
+/*   Updated: 2025/01/24 17:59:02 by emagnani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
